+++ mod-inventory-storage-server/src/main/java/org/folio/services/instance/InstanceService.java
@@ -25,6 +25,7 @@ import static org.folio.validator.NotesValidators.refuseLongNotes;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
+import io.vertx.core.http.HttpClient;
 import io.vertx.core.json.JsonObject;
 import io.vertx.pgclient.PgException;
 import io.vertx.sqlclient.Tuple;
@@ -72,6 +73,7 @@ import org.folio.services.domainevent.InstanceDomainEventPublisher;
 import org.folio.services.reindex.ReindexExportOrchestrator;
 import org.folio.services.sanitizer.Sanitizer;
 import org.folio.services.sanitizer.SanitizerFactory;
+import org.folio.services.setting.SettingsService;
 import org.folio.util.StringUtil;
 import org.folio.validator.CommonValidators;
 import org.folio.validator.NotesValidators;
@@ -93,6 +95,7 @@ public class InstanceService {
   private final InstanceMarcRepository marcRepository;
   private final InstanceRelationshipRepository relationshipRepository;
   private final ConsortiumService consortiumService;
+  private final SettingsService settingsService;
   private final Sanitizer<Instance> sanitizer;
 
   public InstanceService(Context vertxContext, Map<String, String> okapiHeaders) {
@@ -105,8 +108,9 @@ public class InstanceService {
     this.instanceRepository = new InstanceRepository(vertxContext, okapiHeaders);
     this.marcRepository = new InstanceMarcRepository(vertxContext, okapiHeaders);
     this.relationshipRepository = new InstanceRelationshipRepository(vertxContext, okapiHeaders);
-    this.consortiumService = new ConsortiumServiceImpl(vertxContext.owner().createHttpClient(),
+    this.consortiumService = new ConsortiumServiceImpl(vertxContext.get(HttpClient.class.getName()),
       vertxContext.get(ConsortiumDataCache.class.getName()));
+    this.settingsService = new SettingsService(vertxContext, okapiHeaders);
     this.sanitizer = SanitizerFactory.getSanitizer(Instance.class);
   }
 
@@ -269,29 +273,29 @@ public class InstanceService {
   }
 
   private Future<Response> performInstanceUpdate(String id, Instance oldInstance, Instance newInstance) {
-    try {
-      var noChanges = equalsIgnoringMetadata(oldInstance, newInstance);
-      if (noChanges) {
-        return Future.succeededFuture()
-          .map(res -> InstanceStorage.PutInstanceStorageInstancesByInstanceIdResponse.respond204());
-      }
-    } catch (Exception e) {
-      return failedFuture(e);
-    }
-
-    final Promise<Response> putResult = promise();
-    return postgresClient.withTrans(conn -> {
-      Promise<Response> putPromise = putInstance(newInstance, id);
-      return putPromise.future()
-        .compose(response -> linkOrUnlinkSubjects(conn, newInstance, oldInstance)
-          .map(v -> response));
-    }).onComplete(transactionResult -> {
-      if (transactionResult.succeeded()) {
-        putResult.complete(transactionResult.result());
-      } else {
-        putResult.fail(transactionResult.cause());
-      }
-    }).onSuccess(domainEventPublisher.publishUpdated(oldInstance));
+    return settingsService.isOptimizeUpdatesEnabled(okapiHeaders.get(TENANT))
+      .compose(isOptimizeUpdatesEnabled -> {
+        try {
+          if (isOptimizeUpdatesEnabled.booleanValue() && equalsIgnoringMetadata(oldInstance, newInstance)) {
+            return Future.succeededFuture(InstanceStorage.PutInstanceStorageInstancesByInstanceIdResponse.respond204());
+          }
+        } catch (Exception e) {
+          return failedFuture(e);
+        }
+        final Promise<Response> putResult = promise();
+        return postgresClient.withTrans(conn -> {
+          Promise<Response> putPromise = putInstance(newInstance, id);
+          return putPromise.future()
+            .compose(response -> linkOrUnlinkSubjects(conn, newInstance, oldInstance)
+              .map(v -> response));
+        }).onComplete(transactionResult -> {
+          if (transactionResult.succeeded()) {
+            putResult.complete(transactionResult.result());
+          } else {
+            putResult.fail(transactionResult.cause());
+          }
+        }).onSuccess(domainEventPublisher.publishUpdated(oldInstance));
+      });
   }
 
   private Future<Response> postSyncInstance(Conn conn, List<Instance> instances, boolean upsert,