+++ mod-inventory-storage-server/src/main/resources/templates/db_scripts/setEffectiveHoldingsLocation.sql
@@ -1,4 +1,0 @@
-UPDATE ${myuniversity}_${mymodule}.holdings_record
-  SET jsonb = JSONB_SET(holdings_record.jsonb, '{effectiveLocationId}',
-                        TO_JSONB(COALESCE(temporarylocationid, permanentlocationid)))
-WHERE effectivelocationid IS NULL;